% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyFunScript.R
\name{adequat}
\alias{adequat}
\title{Parameter Pearson's chi-squared test. Check over dispersion in a POISON GLM}
\usage{
adequat(model, method = "pearson")
}
\arguments{
\item{model}{A GLM model}

\item{method}{Method to use, 'pearson' is default}
}
\value{
The probability value of the test
}
\description{
Parameter Pearson's chi-squared test. Check over dispersion in a POISON GLM
}
\details{
Parameter Pearson's chi-squared test. H0: le modele est adequat ; H1: le modele n'est pas adequat
}
