% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyFunScript.R
\name{angleMain}
\alias{angleMain}
\title{Main angle}
\usage{
angleMain(angle, angleUnit = c("deg", "rad"))
}
\arguments{
\item{angle}{numeric vector of angle.}

\item{angleUnit}{The unit of \code{angle}. Either "deg", "rad". Default is "deg".}
}
\value{
A matrix of main angle values in radian and in degree units.
}
\description{
\code{angleMain} returns the main angle of an angle value. Main angles range from -pi to pi for radian unit while they range from -180 to 180 for degree unit
}
\seealso{
\code{\link{rad}}, \code{\link{deg}}, \code{\link{slope2angle}}, \code{\link{angle2slope}}
}
