% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyFunScript.R
\name{deg}
\alias{deg}
\title{Radians to degrees}
\usage{
deg(radian)
}
\arguments{
\item{radian}{A vector of degree values to be converted}
}
\value{
Returns a vector of radian values.
}
\description{
Converts angle values from  radian to degree.
}
\examples{
deg(pi/2)  
}
\seealso{
rad
}
