% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyFunScript.R
\name{fibonacci}
\alias{fibonacci}
\title{Fibonacci numbers}
\usage{
fibonacci(n, PrintFib = FALSE, Uo = 0, U1 = 1)
}
\arguments{
\item{n}{Integer ...}

\item{PrintFib}{Logical, indicating if the series should be printed.}

\item{Uo}{Integer ...}

\item{U1}{Integer ...}
}
\value{
Either a real number, result of the function or a vector of al the series.
}
\description{
Generates numbers from Fibonacci serie
}
\examples{
fibonacci(n = 10, PrintFib = TRUE)
fibonacci(n = 10, Uo = 1, U1 = 3, PrintFib = FALSE)
}
\author{
Narcisse Yehouenou
}
