% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyFunScript.R
\name{fiboRate}
\alias{fiboRate}
\title{Fibonacci series}
\usage{
fiboRate(n = 10, PrintSer = FALSE, Uo = 0, U1 = 1)
}
\arguments{
\item{n}{Integer ...}

\item{PrintSer}{Logical, indicating if the series should be printed.}

\item{Uo}{Integer ...}

\item{U1}{Integer ...}
}
\value{
Either a real number, result of the rate of nth and (n-1)th numbers in Fibonacci series.
}
\description{
Computes rates from Fibonacci series
}
\details{
The function returns golden number when Uo = 0, and U1 = 1. Larger n is, more precise the number (result) is.
}
\examples{
fiboRate(n = 100, PrintSer = FALSE, Uo = 0, U1 = 1)
##(1+sqrt(5))/2
fiboRate(n = 10, PrintSer = TRUE, Uo = 0, U1 = 1)
}
\author{
Narcisse Yehouenou
}
