% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyFunScript.R
\name{height}
\alias{height}
\title{Height of tree or vertical object.}
\usage{
height(
  distance,
  vh,
  vb,
  type = c("angle", "slope"),
  angleUnit = c("deg", "rad")
)
}
\arguments{
\item{distance}{A numeric vector of distance between object and the person who measures angle.}

\item{vh, vb}{numeric vector of top angle and down angle respectively.}

\item{type}{The type of 'vh' and 'vb' measures. Either "angle" or "slope". Default is "slope".}

\item{angleUnit}{The unit of 'vh' and 'vb' measures when their are "angle". Either "deg", "rad". Defalt is "deg".}
}
\value{
Returns a vector heights.
}
\description{
Computes the height of tree or vertical object. Allows all both slope and angle measures. No matter the relative position of the persons who measures angle/slope.
}
\examples{
height(10, 80, 17)
height(17, vh = -18, vb = -113)
height(distance = 18, vh = 42, vb = -12, type = "angle", angleUnit = "deg")
height(distance = 18:21, vh = 42:45, vb = -12:-15, type = "angle", angleUnit = "deg")
## Bellow shos warning messages 
height(distance = 18:21, vh = -42:-45, vb = -12:-15, type = "angle", angleUnit = "deg")
}
\author{
Narcisse Yehouenou
}
