% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyFunScript.R
\name{rad}
\alias{rad}
\title{Degrees to radians}
\usage{
rad(degree)
}
\arguments{
\item{degree}{A vector of radian values to be converted}
}
\value{
Returns a vector of radian values.
}
\description{
Converts angle values from  degree to radian.
}
\examples{
rad(180)  
}
\seealso{
\code{\link{deg}}, the the ‘complement’ of \code{rad}
}
