% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyFunScript.R
\name{Vectoriz}
\alias{Vectoriz}
\title{Making factor vectors}
\usage{
Vectoriz(data, binary = FALSE)
}
\arguments{
\item{data}{A data set containing ....}

\item{binary}{Logical indicating if binary data should be considered as factor}
}
\value{
Data frame with all character vectors changed to factor vectors
}
\description{
Changes character vectors to factor vectors
}
